+++ packages/browser-wallet/src/popup/pages/AddWeb3IdCredential/AddWeb3IdCredential.tsx
@@ -63,6 +63,7 @@ export default function AddWeb3IdCredential({ onAllow, onReject }: Props) {
     const network = useAtomValue(networkConfigurationAtom);
 
     const [error, setError] = useState<string>();
+    const [validationComplete, setValidationComplete] = useState<boolean>(false);
 
     const { credential: rawCredential, url, metadataUrl } = state.payload;
     const credential: APIVerifiableCredential = parse(rawCredential);
@@ -121,10 +122,24 @@ export default function AddWeb3IdCredential({ onAllow, onReject }: Props) {
             }
 
             if (missingRequiredAttributeKeys.length > 0) {
-                setError(t('error.attribute', { attributeKeys: missingRequiredAttributeKeys }));
+                setError(t('error.attribute.required', { attributeKeys: missingRequiredAttributeKeys }));
+                setValidationComplete(true);
+                return;
             }
+
+            // Ensure that a credential with more attributes than listed by the schema cannot be added.
+            const schemaAttributes = Object.keys(schema.properties.credentialSubject.properties.attributes.properties);
+            for (const credentialAttribute of Object.keys(credential.credentialSubject.attributes)) {
+                if (!schemaAttributes.includes(credentialAttribute)) {
+                    setError(t('error.attribute.additional', { credentialAttribute, schemaAttributes }));
+                    setValidationComplete(true);
+                    return;
+                }
+            }
+
+            setValidationComplete(true);
         }
-    }, [schema?.properties.credentialSubject.properties.attributes.required]);
+    }, [Boolean(schema)]);
 
     useEffect(() => () => controller.abort(), []);
 
@@ -193,7 +208,7 @@ export default function AddWeb3IdCredential({ onAllow, onReject }: Props) {
         return credentialSubjectId;
     }
 
-    if (web3IdCredentials.loading || storedWeb3IdCredentials.loading) {
+    if (web3IdCredentials.loading || storedWeb3IdCredentials.loading || !validationComplete) {
         return null;
     }
 
@@ -220,7 +235,7 @@ export default function AddWeb3IdCredential({ onAllow, onReject }: Props) {
                         />
                     </>
                 )}
-                <div className="flex justify-center m-t-auto m-b-20">
+                <div className="flex justify-center m-t-auto">
                     <Button width="medium" className="m-r-10" onClick={withClose(onReject)}>
                         {t('reject')}
                     </Button>